Organaize the elements inside the header
I changed the search field appearance. I also removed the
divs that surrounded the profile images and bell icon to
reduce the amount of unecessary code. I changed the header
grid rows to be auto, in order for the padding to be even.
In the profile section I added a greetings span to greet
the user. Set the font size of all the element to suit the
overall look and I addeimported font families.