Finish sidebar style
I added a hover effect to the sidebar menu items. To ahcieve
this effect I removed the padding from the sidebar and set it
directly on the children, this way the list item width is the
same as the sidebar. To provent the unintentional space that
was between the logo and the header, I wrapped them in a div
and set it to flex.