Add main section grid
I also made the projects section a grid. I gave the cards a fixed
width. With repeat auto fit I enabled the cards to 'wrap' into a
single column when the window becomes smaller. I changed the header
to be a fixed size, that way when the cards wrap it will stay the
same size.